IRIS-3963 | merge old tests with new ones
